Update error handling
not sure this is going to work
